Created webview
Webview now works in the electron app and can run via 'npm start' command. I put a manual toggle in main.js whether to load the stable or beta website to test that both work. I would like to give the user a choice which to load and reload appropriately.